Move Page/paginate/PAGE_SIZE to a new services/pagination.py
Nothing about paging a sequence is collection-specific — it was only
defined in services/collection.py because /collection was its first
consumer. Phase 1.1's /banners, /banner_info, and the admin panel's
users table all need the same generic pagination.

Part of #58
