+++ tests/services/test_collection.py
@@ -1,11 +1,11 @@
-"""Tests for the collection service: owned-character lookup + pagination."""
+"""Tests for the collection service: owned-character lookup."""
 
 import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import Session
 
 from ley_shards_bot.models import Base, Character, Player, PlayerCharacter, Rarity
-from ley_shards_bot.services.collection import get_owned_characters, paginate
+from ley_shards_bot.services.collection import get_owned_characters
 
 
 @pytest.fixture
@@ -81,46 +81,3 @@ class TestGetOwnedCharacters:
 
         assert get_owned_characters(session, 1) == []
         assert len(get_owned_characters(session, 2)) == 1
-
-
-class TestPaginate:
-    def test_first_page_of_a_short_list(self):
-        page = paginate([1, 2, 3], page_number=0, page_size=10)
-
-        assert page.items == [1, 2, 3]
-        assert page.total_pages == 1
-        assert page.has_previous is False
-        assert page.has_next is False
-
-    def test_splits_across_pages(self):
-        items = list(range(25))
-
-        first = paginate(items, page_number=0, page_size=10)
-        second = paginate(items, page_number=1, page_size=10)
-        third = paginate(items, page_number=2, page_size=10)
-
-        assert first.items == list(range(10))
-        assert second.items == list(range(10, 20))
-        assert third.items == list(range(20, 25))
-        assert first.total_pages == 3
-        assert first.has_next is True
-        assert first.has_previous is False
-        assert third.has_next is False
-        assert third.has_previous is True
-
-    def test_out_of_range_page_number_clamps_into_range(self):
-        items = list(range(5))
-
-        too_high = paginate(items, page_number=99, page_size=10)
-        too_low = paginate(items, page_number=-5, page_size=10)
-
-        assert too_high.page_number == 0
-        assert too_low.page_number == 0
-
-    def test_empty_list_is_a_single_empty_page(self):
-        page = paginate([], page_number=0, page_size=10)
-
-        assert page.items == []
-        assert page.total_pages == 1
-        assert page.has_previous is False
-        assert page.has_next is False
